+++ src/relation/builder.rs
@@ -153,9 +153,7 @@ impl<RequireInput> MapBuilder<RequireInput> {
 
     /// Add a group by
     pub fn group_by(mut self, expr: Expr) -> Self {
-        self.split = self
-            .split
-            .map_last_reduce(|reduce| reduce.and(Split::group_by(expr)));
+        self.split = self.split.and(Split::group_by(expr.into()).into());
         self
     }
 
@@ -437,6 +435,7 @@ impl<RequireInput> ReduceBuilder<RequireInput> {
     pub fn group_by<E: Into<Expr>>(mut self, expr: E) -> Self {
         self.split = self.split.and(Split::group_by(expr.into()).into());
         self
+
     }
 
     pub fn group_by_iter<I: IntoIterator<Item = Expr>>(self, iter: I) -> Self {